Update deps Fix int32 with int64 since telegram has int64 for users Fix exception handling

+++ src/main/java/com/senderman/lastkatkabot/dbservice/mongodb/MongoChatUserService.java
@@ -69,7 +69,7 @@ public class MongoChatUserService implements ChatUserService {
 
     @Override
     public long getTotalUsers() {
-        return mongoTemplate.findDistinct("userId", ChatUser.class, Integer.class).size();
+        return mongoTemplate.findDistinct("userId", ChatUser.class, Long.class).size();
     }
 
     @Override
